+++ src/pages/addTime/index.tsx
@@ -14,7 +14,7 @@ import { Body, Main, Title, TitleWrapper, Wrapper } from './index.styles';
 
 import { TableSelectedTypes } from './index.types';
 
-import { useGetRoomInfo } from '@/queries/room/useGetRoomInfo';
+import { useGetRoomInfo } from '@/queries/room';
 import { ROUTES } from '@/constants/ROUTES';
 import { Layout } from '@/components/commons/layout';
