--- src/queries/auth/usePostUserInfo.ts
@@ -1,9 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import { postUserInfo } from '@/api/auth';
-import { PostAuthParamsType } from '@/types/auth';
-
-export const usePostUserInfo = () => {
-  return useMutation(({ roomUUID, form }: PostAuthParamsType) =>
-    postUserInfo({ roomUUID, form })
-  );
-};
