--- src/queries/room/useCreateRoom.ts
@@ -1,8 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-import { createRoom } from '@/api/room';
-
-export const useCreateRoom = () => {
-  return useMutation({
-    mutationFn: createRoom,
-  });
-};
